refactor(ocean): scene builders for sky, celestial, islands, lighthouse, ship
Second slice of the split: each builder takes a SceneCtx (scene, the
disposal tracker, the waterline clip plane) and returns the handles
that applyPalette and the render loop mutate, so the orchestrator
keeps owning theme lerps and animation. whiteMat/darkMat stay shared
between lighthouse and ship by riding the lighthouse return value.
Geometry moved verbatim; Ocean3D.tsx is down from ~1800 to ~1040
lines, with water/underwater/scatter/depth still inline since they
are entangled with the per-URL reseeding.
